requirements.txt
# Morning Star Academy - Python Dependencies

# Django and core dependencies
Django==5.2.7
django-tailwind==3.8.0
django-browser-reload==1.12.1

# Environment configuration
python-decouple==3.8
dj-database-url==2.1.0

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Email and async processing
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# PDF generation
weasyprint==60.2
reportlab==4.0.7
Pillow==10.1.0

# Email providers (optional)
sendgrid==6.11.0
mailgun2==1.0.1

# Security and monitoring
django-ratelimit==4.1.0
sentry-sdk[django]==1.38.0

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development tools (optional)
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Documentation (optional)
sphinx==7.2.6
sphinx-rtd-theme==1.3.0